Refaktorerer fetcher ut fra komponent
