+++ src/components/filopplaster/Filopplaster.tsx
@@ -14,44 +14,11 @@ import ReisetilskuddDatovelger from '../dato/ReisetilskuddDatovelger';
 import './Filopplaster.less';
 import env from '../../utils/environment';
 import { logger } from '../../utils/logger';
+import { post } from '../../data/fetcher/fetcher';
 
 interface Props {
   tillatteFiltyper?: string[];
   maxFilstørrelse?: number;
-}
-
-async function fetcher<T>(
-  request: RequestInfo,
-): Promise<HttpResponse<T>> {
-  const response: HttpResponse<T> = await fetch(request);
-  try {
-    response.parsedBody = await response.json();
-  } catch (ex) { logger.error(ex); }
-  if (!response.ok) {
-    throw new Error(response.statusText);
-  }
-  return response;
-}
-
-// eslint-disable-next-line
-async function get<T>(
-  path: string,
-  args: RequestInit = { method: 'get' },
-): Promise<HttpResponse<T>> {
-  return fetcher<T>(new Request(path, args));
-}
-
-async function post<T>(
-  path: string,
-  // eslint-disable-next-line
-  body: any,
-  args: RequestInit = { method: 'post', body: JSON.stringify(body) },
-): Promise<HttpResponse<T>> {
-  return fetcher<T>(new Request(path, args));
-}
-
-interface HttpResponse<T> extends Response {
-  parsedBody?: T;
 }
 
 const Filopplaster: React.FC<Props> = ({ tillatteFiltyper, maxFilstørrelse }) => {
