+++ src/data/fetcher/fetcher.tsx
@@ -1,0 +1,36 @@
+import { logger } from '../../utils/logger';
+
+interface HttpResponse<T> extends Response {
+  parsedBody?: T;
+}
+
+async function fetcher<T>(
+  request: RequestInfo,
+): Promise<HttpResponse<T>> {
+  const response: HttpResponse<T> = await fetch(request);
+  try {
+    response.parsedBody = await response.json();
+  } catch (ex) { logger.error(ex); }
+  if (!response.ok) {
+    throw new Error(response.statusText);
+  }
+  return response;
+}
+
+// eslint-disable-next-line func-names
+export const get = async function<T>(
+  path: string,
+  args: RequestInit = { method: 'get' },
+): Promise<HttpResponse<T>> {
+  return fetcher<T>(new Request(path, args));
+};
+
+// eslint-disable-next-line func-names
+export const post = async function<T>(
+  path: string,
+  // eslint-disable-next-line
+  body: any,
+  args: RequestInit = { method: 'post', body: JSON.stringify(body) },
+): Promise<HttpResponse<T>> {
+  return fetcher<T>(new Request(path, args));
+};
